Rebase to upstream latest master 26 July 2019

+++ plugins/gamma_example/surface_gbc_control.c
@@ -281,7 +281,7 @@ view_draw(	struct spug_view_draw_info *view_draw_info,
 		glUniform1i(timedout_uniform,
 				spug_has_surface_timedout(view_draw_info->id));
 		glUniform1f(opacity_uniform, (current_tile++ == selected_tile) ?
-				abs(60.0 - (float)(frame%120)) / 60.0 :
+				abs((int)(60.0 - (float)(frame%120) / 60.0)) :
 				1.0);
 		glDrawArrays(GL_TRIANGLE_STRIP, 0, 4);
 		glBindTexture(GL_TEXTURE_2D, 0);
@@ -293,7 +293,7 @@ gray_box_draw(void)
 		glUniform1i(gray_uniform, 1);
 		glUniform1i(timedout_uniform, 0);
 		glUniform1f(opacity_uniform, (current_tile++ == selected_tile) ?
-				abs(60.0 - (float)(frame%120)) / 60.0 :
+				abs((int)(60.0 - (float)(frame%120) / 60.0)) :
 				1.0);
 		glDrawArrays(GL_TRIANGLE_STRIP, 0, 4);
 }
